qmltest: rename testimageprovider
