+++ src/qml/test/testimageprovider.h
